cli/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "fairguard-cli"
version = "0.1.0"
description = "FairGuard AI Fairness Firewall CLI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
]

[project.scripts]
fairguard = "fairguard_cli.main:app"

[tool.setuptools.packages.find]
where = ["src"]
